ticket-classifier/requirements.txt
sentence-transformers==2.3.1
transformers==4.36.2
torch>=2.0.0
django==5.0.1
djangorestframework==3.14.0
pandas>=2.0.0
numpy>=1.24.0
requests==2.31.0
python-dotenv==1.0.0
scikit-learn==1.4.0
